# python-flask-server/swagger_server/models/create_service.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util
from swagger_server.models.user import User



class CreateService(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, user_info: User=None, deployment_definition: object=None, service_definition: object=None):  # noqa: E501
        """CreateService - a model defined in Swagger

        :param user_info: The user_info of this CreateService.  # noqa: E501
        :type user_info: User
        :param deployment_definition: The deployment_definition of this CreateService.  # noqa: E501
        :type deployment_definition: object
        :param service_definition: The service_definition of this CreateService.  # noqa: E501
        :type service_definition: object
        """
        self.swagger_types = {
            'user_info': User,
            'deployment_definition': object,
            'service_definition': object
        }

        self.attribute_map = {
            'user_info': 'userInfo',
            'deployment_definition': 'deploymentDefinition',
            'service_definition': 'serviceDefinition'
        }

        self._user_info = user_info
        self._deployment_definition = deployment_definition
        self._service_definition = service_definition

    @classmethod
    def from_dict(cls, dikt) -> 'CreateService':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CreateService of this CreateService.  # noqa: E501
        :rtype: CreateService
        """
        return util.deserialize_model(dikt, cls)

    @property
    def user_info(self) -> User:
        """Gets the user_info of this CreateService.


        :return: The user_info of this CreateService.
        :rtype: User
        """
        return self._user_info

    @user_info.setter
    def user_info(self, user_info: User):
        """Sets the user_info of this CreateService.


        :param user_info: The user_info of this CreateService.
        :type user_info: User
        """

        self._user_info = user_info

    @property
    def deployment_definition(self) -> object:
        """Gets the deployment_definition of this CreateService.


        :return: The deployment_definition of this CreateService.
        :rtype: object
        """
        return self._deployment_definition

    @deployment_definition.setter
    def deployment_definition(self, deployment_definition: object):
        """Sets the deployment_definition of this CreateService.


        :param deployment_definition: The deployment_definition of this CreateService.
        :type deployment_definition: object
        """

        self._deployment_definition = deployment_definition

    @property
    def service_definition(self) -> object:
        """Gets the service_definition of this CreateService.


        :return: The service_definition of this CreateService.
        :rtype: object
        """
        return self._service_definition

    @service_definition.setter
    def service_definition(self, service_definition: object):
        """Sets the service_definition of this CreateService.


        :param service_definition: The service_definition of this CreateService.
        :type service_definition: object
        """

        self._service_definition = service_definition