# python-flask-server/swagger_server/models/service_metadata.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class ServiceMetadata(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, service_id: str=None, ip_address: str=None):  # noqa: E501
        """ServiceMetadata - a model defined in Swagger

        :param service_id: The service_id of this ServiceMetadata.  # noqa: E501
        :type service_id: str
        :param ip_address: The ip_address of this ServiceMetadata.  # noqa: E501
        :type ip_address: str
        """
        self.swagger_types = {
            'service_id': str,
            'ip_address': str
        }

        self.attribute_map = {
            'service_id': 'serviceId',
            'ip_address': 'ipAddress'
        }

        self._service_id = service_id
        self._ip_address = ip_address

    @classmethod
    def from_dict(cls, dikt) -> 'ServiceMetadata':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ServiceMetadata of this ServiceMetadata.  # noqa: E501
        :rtype: ServiceMetadata
        """
        return util.deserialize_model(dikt, cls)

    @property
    def service_id(self) -> str:
        """Gets the service_id of this ServiceMetadata.


        :return: The service_id of this ServiceMetadata.
        :rtype: str
        """
        return self._service_id

    @service_id.setter
    def service_id(self, service_id: str):
        """Sets the service_id of this ServiceMetadata.


        :param service_id: The service_id of this ServiceMetadata.
        :type service_id: str
        """

        self._service_id = service_id

    @property
    def ip_address(self) -> str:
        """Gets the ip_address of this ServiceMetadata.


        :return: The ip_address of this ServiceMetadata.
        :rtype: str
        """
        return self._ip_address

    @ip_address.setter
    def ip_address(self, ip_address: str):
        """Sets the ip_address of this ServiceMetadata.


        :param ip_address: The ip_address of this ServiceMetadata.
        :type ip_address: str
        """

        self._ip_address = ip_address